+++ PerimeterXModule/CustomBehavior/ICustomParametersHandler.cs
@@ -1,0 +1,10 @@
+﻿using System.Web;
+
+namespace PerimeterX.CustomBehavior
+{
+    public interface ICustomParametersHandler
+    {
+        CustomParameters Handle(HttpRequest httpRequest);
+    }
+}
+