+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+
+## [XX.XX.XX] - xxxx-xx-xx
+### Added
+- Added support for custom parameters feature
+
+
 ## [3.2.2] - 2023-08-02
 ### Fixed
 - Added firstPartyEnabled field to advanced blocking response